Gamygdala Engine fixed
